Refs #3344 #3343 #3339. Added integration tests.
